backend/requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.10
python-dotenv==1.1.1
passlib==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
chromadb==1.1.0
python-multipart==0.0.9
pypdf==4.3.1
langchain-nvidia-ai-endpoints==0.1.2
tiktoken==0.11.0